third_party/blink/web_tests/external/wpt/css/css-values/calc-infinity-nan-serialize-length-expected.txt

This is a testharness.js-based test.
[FAIL] 'calc(1 * min(NaN * 2px, NaN * 4em))' as a specified value should serialize as 'calc(NaN * 1px)'.
  assert_equals: 'calc(1 * min(NaN * 2px, NaN * 4em))' and 'calc(NaN * 1px)' should serialize the same in specified values. expected "calc(NaN * 1px)" but got "calc(1 * min(NaN * 1px, NaN * 1em))"
[FAIL] 'calc(1 * clamp(NaN * 2em, NaN * 4px, NaN * 8pt))' as a specified value should serialize as 'clamp(NaN * 1em, NaN * 1px, NaN * 1px)'.
  assert_equals: 'calc(1 * clamp(NaN * 2em, NaN * 4px, NaN * 8pt))' and 'clamp(NaN * 1em, NaN * 1px, NaN * 1px)' should serialize the same in specified values. expected "clamp(NaN * 1em, NaN * 1px, NaN * 1px)" but got "calc(1 * clamp(NaN * 1em, NaN * 1px, NaN * 1px))"
Harness: the test ran to completion.
